+++ src/main/java/bc1/gream/domain/product/service/query/unit/ProductCondition.java
@@ -1,12 +1,10 @@
 package bc1.gream.domain.product.service.query.unit;
 
-import java.math.BigInteger;
-
 public record ProductCondition(
     String brand,
     String name,
-    BigInteger startPrice,
-    BigInteger endPrice
+    Long startPrice,
+    Long endPrice
 ) {
 
 }
